Refactor/config (#142)
* refactor: update ModuleInterface to include configurable settings for enabled state


* refactor: simplify module filtering logic in ModuleManager by using enabled() method


* refactor: update modules to use config properties instead of ConfigManager for better modularity


* refactor: change enabled property in module configs from val to var for better configurability


* refactor: enhance module configuration by introducing AllConfigs data class and loading settings in ModuleManager


* refactor: replace ConfigManager with InvUnloadModule.Config for improved configuration access


* refactor: add JsonIgnore annotation to enabled property for improved serialization control


* refactor: update ConfigManager and ModuleManager to improve configuration loading and saving


* refactor: remove unnecessary suppression annotation in ModuleManager for cleaner code


---------
