+++ .idea/dictionaries/project.xml
@@ -17,7 +17,9 @@
       <w>eldoria</w>
       <w>enderman</w>
       <w>gmazzo</w>
+      <w>invs</w>
       <w>invsearch</w>
+      <w>invu</w>
       <w>invunload</w>
       <w>levelup</w>
       <w>luckperms</w>
@@ -25,9 +27,11 @@
       <w>mojang</w>
       <w>netherite</w>
       <w>newmoon</w>
+      <w>searchinv</w>
       <w>shulker</w>
       <w>subcmd</w>
       <w>triumphteam</w>
+      <w>unloadinv</w>
       <w>useable</w>
       <w>vanillaplus</w>
       <w>waystone</w>
