+++ src/main/kotlin/org/xodium/vanillaplus/interfaces/ModuleInterface.kt
@@ -10,7 +10,21 @@ import org.bukkit.permissions.Permission
 import org.xodium.vanillaplus.data.CommandData
 
 /** Represents a contract for a module within the system. */
-interface ModuleInterface : Listener {
+interface ModuleInterface<out T : ModuleInterface.Config> : Listener {
+    /**
+     * Represents the configuration settings for a module.
+     * @property enabled Indicates whether the module is enabled or not.
+     */
+    interface Config {
+        var enabled: Boolean
+    }
+
+    /**
+     * Retrieves the configuration for this module.
+     * @return A [Config] object containing the module's configuration settings.
+     */
+    val config: T
+
     /**
      * Determines if this module is currently enabled.
      * @return `true` if the module is enabled, `false` otherwise.
