+++ src/main/kotlin/org/xodium/vanillaplus/managers/ConfigManager.kt
@@ -8,11 +8,12 @@ package org.xodium.vanillaplus.managers
 import com.fasterxml.jackson.annotation.JsonAutoDetect
 import com.fasterxml.jackson.annotation.PropertyAccessor
 import com.fasterxml.jackson.databind.DeserializationFeature
+import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.SerializationFeature
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
-import org.xodium.vanillaplus.data.ConfigData
+import org.xodium.vanillaplus.interfaces.ModuleInterface
 import java.io.IOException
 import java.nio.file.Files
 import java.nio.file.StandardOpenOption
@@ -20,40 +21,39 @@ import java.nio.file.StandardOpenOption
 /** Represents the config manager within the system. */
 object ConfigManager {
     private val configPath = instance.dataFolder.toPath().resolve("config.json")
-    private val objectMapper = jacksonObjectMapper()
+    internal val objectMapper = jacksonObjectMapper()
         .registerModules(JavaTimeModule())
         .disable(SerializationFeature.FAIL_ON_EMPTY_BEANS)
         .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
         .setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.ANY)
 
-    @Volatile
-    var data: ConfigData = ConfigData()
-        private set
-
     /**
-     * Initializes the configuration by loading settings from the config file.
+     * Loads settings from the config file.
      * @param silent If true, suppresses logging messages during loading.
+     * @return A JsonNode representing the configuration, or null on failure or if file doesn't exist.
      */
-    fun load(silent: Boolean = false) {
+    fun load(silent: Boolean = false): JsonNode? {
         try {
-            data = ConfigData()
             if (Files.exists(configPath)) {
                 if (!silent) instance.logger.info("Config: Loading settings.")
-                data = objectMapper
-                    .readerForUpdating(data)
-                    .readValue(Files.readString(configPath))
+                val node = objectMapper.readTree(Files.readString(configPath))
                 if (!silent) instance.logger.info("Config: Settings loaded successfully.")
+                return node
             } else {
-                instance.logger.info("Config: No config file found, creating new config.")
+                if (!silent) instance.logger.info("Config: No config file found, creating new one.")
+                return null
             }
-            save()
         } catch (e: IOException) {
             instance.logger.severe("Config: Failed to load config file: ${e.printStackTrace()}")
+            return null
         }
     }
 
-    /** Saves the current settings to the config file. */
-    private fun save() {
+    /**
+     * Saves the current settings to the config file.
+     * @param data The map of configuration data to save.
+     */
+    fun save(data: Map<String, ModuleInterface.Config>) {
         try {
             instance.logger.info("Config: Saving settings.")
             Files.createDirectories(configPath.parent)
