+++ src/main/kotlin/org/xodium/vanillaplus/modules/ChatModule.kt
@@ -23,15 +23,19 @@ import org.xodium.vanillaplus.VanillaPlus.Companion.PREFIX
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.data.CommandData
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.managers.ConfigManager
+import org.xodium.vanillaplus.utils.ExtUtils.clickRunCmd
 import org.xodium.vanillaplus.utils.ExtUtils.mm
 import org.xodium.vanillaplus.utils.FmtUtils.fireFmt
+import org.xodium.vanillaplus.utils.FmtUtils.mangoFmt
+import org.xodium.vanillaplus.utils.FmtUtils.skylineFmt
 import org.xodium.vanillaplus.utils.SkinUtils.faceToMM
 import org.xodium.vanillaplus.utils.Utils
 import java.util.concurrent.CompletableFuture
 
-class ChatModule : ModuleInterface {
-    override fun enabled(): Boolean = ConfigManager.data.chatModule.enabled
+class ChatModule : ModuleInterface<ChatModule.Config> {
+    override val config: Config = Config()
+
+    override fun enabled(): Boolean = config.enabled
 
     @Suppress("UnstableApiUsage")
     override fun cmds(): CommandData? {
@@ -86,9 +90,9 @@ class ChatModule : ModuleInterface {
 
     @EventHandler(priority = EventPriority.HIGHEST)
     fun on(event: AsyncChatEvent) {
-        if (!enabled()) return
+        if (!config.enabled) return
         event.renderer { source, displayName, message, _ ->
-            ConfigManager.data.chatModule.chatFormat.mm(
+            config.chatFormat.mm(
                 Placeholder.component(
                     "player",
                     displayName
@@ -102,7 +106,7 @@ class ChatModule : ModuleInterface {
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     fun on(event: PlayerJoinEvent) {
-        if (!enabled()) return
+        if (!config.enabled) return
 
         event.joinMessage(null)
 
@@ -111,7 +115,7 @@ class ChatModule : ModuleInterface {
             .filter { it.uniqueId != player.uniqueId }
             .forEach {
                 it.sendMessage(
-                    ConfigManager.data.chatModule.joinMessage.mm(
+                    config.joinMessage.mm(
                         Placeholder.component("player", player.displayName())
                     )
                 )
@@ -120,7 +124,7 @@ class ChatModule : ModuleInterface {
         val faceLines = player.faceToMM().lines()
         var imageIndex = 1
         val welcomeText =
-            Regex("<image>").replace(ConfigManager.data.chatModule.welcomeText) { "<image${imageIndex++}>" }
+            Regex("<image>").replace(config.welcomeText) { "<image${imageIndex++}>" }
         val imageResolvers = faceLines.mapIndexed { i, line -> Placeholder.component("image${i + 1}", line.mm()) }
         val playerComponent = player
             .displayName()
@@ -137,7 +141,7 @@ class ChatModule : ModuleInterface {
 
     @EventHandler(priority = EventPriority.MONITOR, ignoreCancelled = true)
     fun on(event: PlayerQuitEvent) {
-        if (!enabled()) return
+        if (!config.enabled) return
 
         event.quitMessage(null)
 
@@ -146,7 +150,7 @@ class ChatModule : ModuleInterface {
             .filter { it.uniqueId != player.uniqueId }
             .forEach {
                 it.sendMessage(
-                    ConfigManager.data.chatModule.quitMessage.mm(
+                    config.quitMessage.mm(
                         Placeholder.component("player", player.displayName())
                     )
                 )
@@ -161,7 +165,7 @@ class ChatModule : ModuleInterface {
      */
     private fun whisper(sender: Player, target: Player, message: String) {
         sender.sendMessage(
-            ConfigManager.data.chatModule.whisperToFormat.mm(
+            config.whisperToFormat.mm(
                 Placeholder.component(
                     "player",
                     target.displayName()
@@ -173,7 +177,7 @@ class ChatModule : ModuleInterface {
         )
 
         target.sendMessage(
-            ConfigManager.data.chatModule.whisperFromFormat.mm(
+            config.whisperFromFormat.mm(
                 Placeholder.component(
                     "player",
                     sender.displayName()
@@ -184,4 +188,28 @@ class ChatModule : ModuleInterface {
             )
         )
     }
+
+    data class Config(
+        override var enabled: Boolean = true,
+        var chatFormat: String = "<player> <reset>${"›".mangoFmt(true)} <message>",
+        var welcomeText: String =
+            """
+        ${"]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)} ${"Welcome".fireFmt()} <player>
+        <image>${"⯈".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)} ${"Check out".fireFmt()}<gray>: ${
+                "/rules".clickRunCmd(Utils.cmdHover).skylineFmt()
+            } <gray>🟅 ${"/guide".clickRunCmd(Utils.cmdHover).skylineFmt()}
+        <image>${"⯈".mangoFmt(true)}
+        <image>${"⯈".mangoFmt(true)}
+        ${"]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[=]|[".mangoFmt(true)}
+        """.trimIndent(),
+        var joinMessage: String = "<green>➕<reset> ${"›".mangoFmt(true)} <player>",
+        var quitMessage: String = "<red>➖<reset> ${"›".mangoFmt(true)} <player>",
+        var whisperToFormat: String = "${"You".skylineFmt()} ${"➛".mangoFmt(true)} <player> <reset>${"›".mangoFmt(true)} <message>",
+        var whisperFromFormat: String = "<player> <reset>${"➛".mangoFmt(true)} ${"You".skylineFmt()} ${"›".mangoFmt(true)} <message>",
+    ) : ModuleInterface.Config
 }
