+++ src/main/kotlin/org/xodium/vanillaplus/modules/InvSearchModule.kt
@@ -25,7 +25,6 @@ import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.data.CommandData
 import org.xodium.vanillaplus.interfaces.ModuleInterface
 import org.xodium.vanillaplus.managers.ChestAccessManager
-import org.xodium.vanillaplus.managers.ConfigManager
 import org.xodium.vanillaplus.managers.CooldownManager
 import org.xodium.vanillaplus.utils.ExtUtils.mm
 import org.xodium.vanillaplus.utils.FmtUtils.fireFmt
@@ -35,8 +34,10 @@ import org.xodium.vanillaplus.utils.Utils
 import java.util.concurrent.CompletableFuture
 
 /** Represents a module handling inv-search mechanics within the system. */
-class InvSearchModule : ModuleInterface {
-    override fun enabled(): Boolean = ConfigManager.data.invSearchModule.enabled
+class InvSearchModule : ModuleInterface<InvSearchModule.Config> {
+    override val config: Config = Config()
+
+    override fun enabled(): Boolean = config.enabled
 
     override fun cmds(): CommandData? {
         return CommandData(
@@ -99,14 +100,14 @@ class InvSearchModule : ModuleInterface {
      */
     private fun search(player: Player, material: Material) {
         val cooldownKey = NamespacedKey(instance, "invsearch_cooldown")
-        val cooldownDuration = ConfigManager.data.invSearchModule.cooldown
+        val cooldownDuration = config.cooldown
         if (CooldownManager.isOnCooldown(player, cooldownKey, cooldownDuration)) {
             return player.sendActionBar("You must wait before using this again.".fireFmt().mm())
         }
         CooldownManager.setCooldown(player, cooldownKey, System.currentTimeMillis())
 
         val deniedChestKey = NamespacedKey(instance, "denied_chest")
-        val chests = Utils.findBlocksInRadius(player.location, ConfigManager.data.invSearchModule.searchRadius)
+        val chests = Utils.findBlocksInRadius(player.location, config.searchRadius)
             .filter { ChestAccessManager.isAllowed(player, deniedChestKey, it) }
         if (chests.isEmpty()) {
             return player.sendActionBar("No usable chests found for ${"$material".roseFmt()}".fireFmt().mm())
@@ -220,4 +221,10 @@ class InvSearchModule : ModuleInterface {
             }
         }
     }
+
+    data class Config(
+        override var enabled: Boolean = true,
+        var cooldown: Long = 1L * 1000L,
+        var searchRadius: Int = 5,
+    ) : ModuleInterface.Config
 }
