+++ src/main/kotlin/org/xodium/vanillaplus/modules/MotdModule.kt
@@ -9,16 +9,27 @@ import org.bukkit.event.EventHandler
 import org.bukkit.event.EventPriority
 import org.bukkit.event.server.ServerListPingEvent
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.managers.ConfigManager
 import org.xodium.vanillaplus.utils.ExtUtils.mm
+import org.xodium.vanillaplus.utils.FmtUtils.fireFmt
+import org.xodium.vanillaplus.utils.FmtUtils.mangoFmt
 
 /** Represents a module handling MOTD mechanics within the system. */
-class MotdModule : ModuleInterface {
-    override fun enabled(): Boolean = ConfigManager.data.motdModule.enabled
+class MotdModule : ModuleInterface<MotdModule.Config> {
+    override val config: Config = Config()
+
+    override fun enabled(): Boolean = config.enabled
 
     @EventHandler(priority = EventPriority.MONITOR)
     fun on(event: ServerListPingEvent) {
         if (!enabled()) return
-        event.motd(ConfigManager.data.motdModule.motd.joinToString("\n").mm())
+        event.motd(config.motd.joinToString("\n").mm())
     }
+
+    data class Config(
+        override var enabled: Boolean = true,
+        val motd: List<String> = listOf(
+            "<b>Ultimate Private SMP</b>".fireFmt(),
+            "<b>➤ WELCOME BACK LADS!</b>".mangoFmt()
+        )
+    ) : ModuleInterface.Config
 }
