+++ src/main/kotlin/org/xodium/vanillaplus/modules/RecipiesModule.kt
@@ -11,11 +11,12 @@ import org.bukkit.event.EventPriority
 import org.bukkit.event.player.PlayerJoinEvent
 import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.managers.ConfigManager
 
 /** Represents a module handling recipe mechanics within the system. */
-class RecipiesModule : ModuleInterface {
-    override fun enabled(): Boolean = ConfigManager.data.recipiesModule.enabled
+class RecipiesModule : ModuleInterface<RecipiesModule.Config> {
+    override val config: Config = Config()
+
+    override fun enabled(): Boolean = config.enabled
 
     @EventHandler(priority = EventPriority.MONITOR)
     fun on(event: PlayerJoinEvent) {
@@ -24,4 +25,8 @@ class RecipiesModule : ModuleInterface {
             instance.server.recipeIterator().asSequence().filterIsInstance<Keyed>().map { it.key }.toList()
         )
     }
+
+    data class Config(
+        override var enabled: Boolean = true
+    ) : ModuleInterface.Config
 }
