+++ src/main/kotlin/org/xodium/vanillaplus/modules/TrowelModule.kt
@@ -21,14 +21,15 @@ import org.xodium.vanillaplus.VanillaPlus.Companion.instance
 import org.xodium.vanillaplus.data.CommandData
 import org.xodium.vanillaplus.data.TrowelStateData
 import org.xodium.vanillaplus.interfaces.ModuleInterface
-import org.xodium.vanillaplus.managers.ConfigManager
 import org.xodium.vanillaplus.utils.ExtUtils.mm
 import org.xodium.vanillaplus.utils.FmtUtils.fireFmt
 import org.xodium.vanillaplus.utils.Utils
 
 /** Represents a module handling trowel mechanics within the system. */
-class TrowelModule : ModuleInterface {
-    override fun enabled(): Boolean = ConfigManager.data.trowelModule.enabled
+class TrowelModule : ModuleInterface<TrowelModule.Config> {
+    override val config: Config = Config()
+
+    override fun enabled(): Boolean = config.enabled
 
     @Suppress("UnstableApiUsage")
     override fun cmds(): CommandData? {
@@ -93,4 +94,8 @@ class TrowelModule : ModuleInterface {
         val msg = if (enabled) "Trowel: <green>enabled" else "Trowel: <red>disabled"
         player.sendActionBar(msg.fireFmt().mm())
     }
+
+    data class Config(
+        override var enabled: Boolean = true
+    ) : ModuleInterface.Config
 }
